+++ src/main/java/com/inq/wishhair/wesharewishhair/user/controller/dto/response/MyPageResponse.java
@@ -1,6 +1,6 @@
 package com.inq.wishhair.wesharewishhair.user.controller.dto.response;
 
-import com.inq.wishhair.wesharewishhair.user.domain.point.domain.PointHistory;
+import com.inq.wishhair.wesharewishhair.user.domain.point.PointHistory;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.enums.Sex;
 import lombok.AllArgsConstructor;
