+++ src/main/java/com/inq/wishhair/wesharewishhair/user/domain/point/PointRepository.java
@@ -1,4 +1,4 @@
-package com.inq.wishhair.wesharewishhair.user.domain.point.domain;
+package com.inq.wishhair.wesharewishhair.user.domain.point;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
